dd
dd

+++ include/collision_particles_cube.h
@@ -13,12 +13,12 @@ public:
 	collision_particles_cube(QString& _name, modeler* _md, particle_system *_ps, cube *_c, tContactModel _tcm);
 	virtual ~collision_particles_cube();
 
-	virtual bool collid(float dt);
+	virtual bool collid(double dt);
 	virtual bool cuCollid();
-	virtual bool collid_with_particle(unsigned int i, float dt);
+	virtual bool collid_with_particle(unsigned int i, double dt);
 
 private:
-	bool HMCModel(unsigned inti, float dt);
+	bool HMCModel(unsigned inti, double dt);
 	particle_system *ps;
 	cube *cu;
 };
