+++ include/simulation/dembd_simulation.h
@@ -14,16 +14,16 @@ public:
 
 	virtual bool initialize(bool isCpu);
 
-	bool saveResult(float t, unsigned int part);
+	bool saveResult(double t, unsigned int part);
 
 public slots:
 	virtual bool cpuRun();
 	virtual bool gpuRun();
 
 private:
-	void predictionStep(float dt);
-	void collision_dembd(float dt);
-	void correctionStep(float dt);
+	void predictionStep(double dt);
+	void collision_dembd(double dt);
+	void correctionStep(double dt);
 	dem_simulation *dem;
 	mbd_simulation *mbd;
 };
